remove configuration dependency on vscode
